pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-business-workflows"
version = "0.1.0"
description = "An opinionated MCP server for business workflows — triage, routing, human review, and operational actions"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    "mcp>=1.0.0",
    "fastapi>=0.111.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "httpx>=0.27.0",
    "uvicorn[standard]>=0.30.0",
    "structlog>=24.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-httpx>=0.30.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "google-genai>=1.0.0",
    "anthropic>=0.30.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
mcp-business-workflows = "mcp_business_workflows.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_business_workflows"]

[tool.ruff]
src = ["src"]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
files = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
